Minor changes for protocol v3

+++ api/src/main/java/org/cloudburstmc/api/entity/misc/AreaEffectCloud.java
@@ -36,7 +36,7 @@ public interface AreaEffectCloud extends Entity {
 
     long getSpawnTime();
 
-    void setSpawnTime(long spawnTime);
+    void setSpawnTime(int spawnTime);
 
     int getDuration();
